requirements.txt
keyboard ~= 0.13.5
monitorcontrol ~= 3.1.0
pillow~=10.2.0
pystray ~= 0.19.5
